+++ test/typescript/typescript-tests.ts
@@ -436,11 +436,11 @@ test("action decorator (typescript)", function(t) {
 	t.equal(store1.add(1, 1), 4);
 
 	t.deepEqual(normalizeSpyEvents(events),	[
-		{ arguments: [ 3, 4 ], name: "add", spyReportStart: true, target: store1, type: "action" },
+		{ arguments: [ 3, 4 ], name: "add", spyReportStart: true, object: store1, type: "action" },
 		{ spyReportEnd: true },
-		{ arguments: [ 2, 2 ], name: "add", spyReportStart: true, target: store2, type: "action" },
+		{ arguments: [ 2, 2 ], name: "add", spyReportStart: true, object: store2, type: "action" },
 		{ spyReportEnd: true },
-		{ arguments: [ 1, 1 ], name: "add", spyReportStart: true, target: store1, type: "action" },
+		{ arguments: [ 1, 1 ], name: "add", spyReportStart: true, object: store1, type: "action" },
 		{ spyReportEnd: true }
 	]);
 
@@ -467,11 +467,11 @@ test("custom action decorator (typescript)", function(t) {
 	t.equal(store1.add(1, 1), 4);
 
 	t.deepEqual(normalizeSpyEvents(events),	[
-		{ arguments: [ 3, 4 ], name: "zoem zoem", spyReportStart: true, target: store1, type: "action" },
+		{ arguments: [ 3, 4 ], name: "zoem zoem", spyReportStart: true, object: store1, type: "action" },
 		{ spyReportEnd: true },
-		{ arguments: [ 2, 2 ], name: "zoem zoem", spyReportStart: true, target: store2, type: "action" },
+		{ arguments: [ 2, 2 ], name: "zoem zoem", spyReportStart: true, object: store2, type: "action" },
 		{ spyReportEnd: true },
-		{ arguments: [ 1, 1 ], name: "zoem zoem", spyReportStart: true, target: store1, type: "action" },
+		{ arguments: [ 1, 1 ], name: "zoem zoem", spyReportStart: true, object: store1, type: "action" },
 		{ spyReportEnd: true }
 	]);
 
@@ -499,11 +499,11 @@ test("action decorator on field (typescript)", function(t) {
 	t.equal(store1.add(2, 2), 8);
 
 	t.deepEqual(normalizeSpyEvents(events),	[
-		{ arguments: [ 3, 4 ], name: "add", spyReportStart: true, target: store1, type: "action" },
+		{ arguments: [ 3, 4 ], name: "add", spyReportStart: true, object: store1, type: "action" },
 		{ spyReportEnd: true },
-		{ arguments: [ 4, 5 ], name: "add", spyReportStart: true, target: store2, type: "action" },
+		{ arguments: [ 4, 5 ], name: "add", spyReportStart: true, object: store2, type: "action" },
 		{ spyReportEnd: true },
-		{ arguments: [ 2, 2 ], name: "add", spyReportStart: true, target: store1, type: "action" },
+		{ arguments: [ 2, 2 ], name: "add", spyReportStart: true, object: store1, type: "action" },
 		{ spyReportEnd: true }
 	]);
 
@@ -531,11 +531,11 @@ test("custom action decorator on field (typescript)", function(t) {
 	t.equal(store1.add(2, 2), 8);
 
 	t.deepEqual(normalizeSpyEvents(events),	[
-		{ arguments: [ 3, 4 ], name: "zoem zoem", spyReportStart: true, target: store1, type: "action" },
+		{ arguments: [ 3, 4 ], name: "zoem zoem", spyReportStart: true, object: store1, type: "action" },
 		{ spyReportEnd: true },
-		{ arguments: [ 4, 5 ], name: "zoem zoem", spyReportStart: true, target: store2, type: "action" },
+		{ arguments: [ 4, 5 ], name: "zoem zoem", spyReportStart: true, object: store2, type: "action" },
 		{ spyReportEnd: true },
-		{ arguments: [ 2, 2 ], name: "zoem zoem", spyReportStart: true, target: store1, type: "action" },
+		{ arguments: [ 2, 2 ], name: "zoem zoem", spyReportStart: true, object: store1, type: "action" },
 		{ spyReportEnd: true }
 	]);